# frontend/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoriaRequestDto": {
        "properties": {
          "descripcion": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "nombre"
        ],
        "type": "object"
      },
      "CategoriaResponseDto": {
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ComplementoRequestDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "insumoAsociadoId": {
            "format": "uuid",
            "type": "string"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "precioAdicional": {
            "exclusiveMinimum": false,
            "minimum": 0,
            "type": "number"
          },
          "tipoComplemento": {
            "enum": [
              "BEBIDA",
              "HELADO",
              "CERVEZA",
              "JUGO",
              "SALSAS",
              "OTROS"
            ],
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "precioAdicional",
          "tipoComplemento"
        ],
        "type": "object"
      },
      "ComplementoResponseDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "insumoAsociadoId": {
            "format": "uuid",
            "type": "string"
          },
          "insumoAsociadoNombre": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "precioAdicional": {
            "type": "number"
          },
          "tipoComplemento": {
            "enum": [
              "BEBIDA",
              "HELADO",
              "CERVEZA",
              "JUGO",
              "SALSAS",
              "OTROS"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "InsumoRequestDto": {
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "stockMinimo": {
            "exclusiveMinimum": false,
            "minimum": 0,
            "type": "number"
          },
          "tipoInsumo": {
            "enum": [
              "COCIDO",
              "NO_COCIDO"
            ],
            "type": "string"
          },
          "unidadMedida": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "tipoInsumo",
          "unidadMedida"
        ],
        "type": "object"
      },
      "InsumoResponseDto": {
        "properties": {
          "bajoMinimo": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "stockActual": {
            "type": "number"
          },
          "stockMinimo": {
            "type": "number"
          },
          "tipoInsumo": {
            "enum": [
              "COCIDO",
              "NO_COCIDO"
            ],
            "type": "string"
          },
          "unidadMedida": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PageResponseDtoInsumoResponseDto": {
        "properties": {
          "contenido": {
            "items": {
              "$ref": "#/components/schemas/InsumoResponseDto"
            },
            "type": "array"
          },
          "pagina": {
            "format": "int32",
            "type": "integer"
          },
          "primera": {
            "type": "boolean"
          },
          "tamano": {
            "format": "int32",
            "type": "integer"
          },
          "totalElementos": {
            "format": "int64",
            "type": "integer"
          },
          "totalPaginas": {
            "format": "int32",
            "type": "integer"
          },
          "ultima": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "Pageable": {
        "properties": {
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API de POS, inventario, comandas, cocina, despacho y caja del sistema de gestion gastronomica Chaquena.",
    "title": "Backend Logistica - Chaquena",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/categorias": {
      "get": {
        "operationId": "listarCategorias",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoriaResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Listar categorias de la carta",
        "tags": [
          "Catalogo - Categorias"
        ]
      },
      "post": {
        "operationId": "crearCategoria",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Crear categoria",
        "tags": [
          "Catalogo - Categorias"
        ]
      }
    },
    "/api/v1/categorias/{id}": {
      "delete": {
        "operationId": "eliminar",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Eliminar categoria sin platillos asociados",
        "tags": [
          "Catalogo - Categorias"
        ]
      },
      "get": {
        "operationId": "obtenerCategoria",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Obtener categoria por id",
        "tags": [
          "Catalogo - Categorias"
        ]
      },
      "put": {
        "operationId": "actualizarCategoria",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Actualizar categoria",
        "tags": [
          "Catalogo - Categorias"
        ]
      }
    },
...
    "/api/v1/complementos": {
      "get": {
        "operationId": "listarComplementos",
        "parameters": [
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "enum": [
                "BEBIDA",
                "HELADO",
                "CERVEZA",
                "JUGO",
                "SALSAS",
                "OTROS"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "soloActivos",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComplementoResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Listar complementos, opcionalmente por tipo",
        "tags": [
          "Catalogo - Complementos"
        ]
      },
      "post": {
        "operationId": "crearComplemento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplementoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplementoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Crear complemento",
        "tags": [
          "Catalogo - Complementos"
        ]
      }
    },
    "/api/v1/complementos/{id}": {
      "get": {
        "operationId": "obtenerComplemento",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplementoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Obtener complemento por id",
        "tags": [
          "Catalogo - Complementos"
        ]
      },
      "put": {
        "operationId": "actualizarComplemento",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplementoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplementoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Actualizar complemento",
        "tags": [
          "Catalogo - Complementos"
        ]
      }
    },
    "/api/v1/complementos/{id}/activo": {
      "patch": {
        "operationId": "cambiarActivoComplemento",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activo",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplementoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Habilitar o deshabilitar complemento",
        "tags": [
          "Catalogo - Complementos"
        ]
      }
    },
...
    "/api/v1/insumos": {
      "get": {
        "operationId": "buscarInsumos",
        "parameters": [
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "enum": [
                "COCIDO",
                "NO_COCIDO"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "bajoMinimo",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "pageable",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoInsumoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Listar insumos con filtros por tipo, texto y bajo minimo",
        "tags": [
          "Inventario - Insumos"
        ]
      },
      "post": {
        "operationId": "crearInsumo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsumoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsumoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Crear insumo (nace con stock cero)",
        "tags": [
          "Inventario - Insumos"
        ]
      }
    },
...
    "/api/v1/insumos/{id}": {
      "get": {
        "operationId": "obtenerInsumo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsumoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Obtener insumo por id",
        "tags": [
          "Inventario - Insumos"
        ]
      },
      "put": {
        "operationId": "actualizarInsumo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsumoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsumoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Actualizar datos maestros del insumo (no mueve stock)",
        "tags": [
          "Inventario - Insumos"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}